# Conflicts:
#	src/main/java/com/convenientservices/web/repositories/PointOfServicesRepository.java
#	src/main/java/com/convenientservices/web/services/PointOfServiceServices.java

+++ src/main/java/com/convenientservices/web/services/CityService.java
@@ -1,12 +1,12 @@
 package com.convenientservices.web.services;
 
-import com.convenientservices.web.Exceptions.CityNotFoundException;
+import com.convenientservices.web.exceptions.CityNotFoundException;
 import com.convenientservices.web.entities.City;
 
 import java.util.List;
 
 public interface CityService {
-    City findById (Long id) throws Exception;
+    City findById (Integer id) throws Exception;
     List<City> findAll();
     City save(City city);
     City findByName(String name) throws CityNotFoundException;
